+++ src/test/java/tfw/immutable/ila/doubleila/DoubleIlaFromCastFloatIlaTest.java
@@ -24,7 +24,7 @@ class DoubleIlaFromCastFloatIlaTest {
         FloatIla ila = FloatIlaFromArray.create(array);
         DoubleIla targetIla = DoubleIlaFromArray.create(target);
         DoubleIla actualIla = DoubleIlaFromCastFloatIla.create(ila);
-        final double epsilon = (double) 0.0;
+        final double epsilon = 0.0;
         DoubleIlaCheck.checkAll(
                 targetIla,
                 actualIla,
