+++ src/test/java/tfw/immutable/ila/floatila/FloatIlaFromCastCharIlaTest.java
@@ -24,7 +24,7 @@ class FloatIlaFromCastCharIlaTest {
         CharIla ila = CharIlaFromArray.create(array);
         FloatIla targetIla = FloatIlaFromArray.create(target);
         FloatIla actualIla = FloatIlaFromCastCharIla.create(ila);
-        final float epsilon = (float) 0.0;
+        final float epsilon = 0.0f;
         FloatIlaCheck.checkAll(
                 targetIla,
                 actualIla,
