+++ src/test/java/tfw/immutable/ila/floatila/FloatIlaInvertTest.java
@@ -22,7 +22,7 @@ class FloatIlaInvertTest {
         FloatIla ila = FloatIlaFromArray.create(array);
         FloatIla targetIla = FloatIlaFromArray.create(target);
         FloatIla actualIla = FloatIlaInvert.create(ila);
-        final float epsilon = (float) 0.0;
+        final float epsilon = 0.0f;
         FloatIlaCheck.checkAll(
                 targetIla,
                 actualIla,
