+++ src/test/java/tfw/immutable/ila/longila/LongIlaFromCastFloatIlaTest.java
@@ -24,7 +24,7 @@ class LongIlaFromCastFloatIlaTest {
         FloatIla ila = FloatIlaFromArray.create(array);
         LongIla targetIla = LongIlaFromArray.create(target);
         LongIla actualIla = LongIlaFromCastFloatIla.create(ila);
-        final long epsilon = (long) 0.0;
+        final long epsilon = 0L;
         LongIlaCheck.checkAll(
                 targetIla,
                 actualIla,
