+++ src/test/java/tfw/immutable/ila/longila/LongIlaSubtractTest.java
@@ -19,13 +19,13 @@ class LongIlaSubtractTest {
         for (int ii = 0; ii < leftArray.length; ++ii) {
             leftArray[ii] = random.nextLong();
             rightArray[ii] = random.nextLong();
-            array[ii] = (long) (leftArray[ii] - rightArray[ii]);
+            array[ii] = (leftArray[ii] - rightArray[ii]);
         }
         LongIla leftIla = LongIlaFromArray.create(leftArray);
         LongIla rightIla = LongIlaFromArray.create(rightArray);
         LongIla targetIla = LongIlaFromArray.create(array);
         LongIla actualIla = LongIlaSubtract.create(leftIla, rightIla);
-        final long epsilon = (long) 0.0;
+        final long epsilon = 0L;
         LongIlaCheck.checkAll(
                 targetIla,
                 actualIla,
